% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_means.R
\name{chunk_means}
\alias{chunk_means}
\title{FUNCTION_TITLE}
\usage{
chunk_means(chunk_time, job_chunks, chunk_complete)
}
\arguments{
\item{chunk_time}{PARAM_DESCRIPTION}

\item{job_chunks}{PARAM_DESCRIPTION}

\item{chunk_complete}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
