% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stan_tail.R
\name{clean_stan_tail}
\alias{clean_stan_tail}
\title{FUNCTION_TITLE}
\usage{
clean_stan_tail(value)
}
\arguments{
\item{value}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
