% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibble.R
\name{qibble}
\alias{qibble}
\title{FUNCTION_TITLE}
\usage{
qibble(workDir = "qapply", tag = "stanrun")
}
\arguments{
\item{workDir}{PARAM_DESCRIPTION, Default: 'qapply'}

\item{tag}{PARAM_DESCRIPTION, Default: 'stanrun'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
