% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_jobs.R
\name{current_jobs}
\alias{current_jobs}
\title{FUNCTION_TITLE}
\usage{
current_jobs(d)
}
\arguments{
\item{d}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
