% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete.R
\name{get_complete}
\alias{get_complete}
\title{FUNCTION_TITLE}
\usage{
get_complete(workDir = "qapply", tag = "stanrun")
}
\arguments{
\item{workDir}{PARAM_DESCRIPTION, Default: 'qapply'}

\item{tag}{PARAM_DESCRIPTION, Default: 'stanrun'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
