% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counter.R
\name{get_counter}
\alias{get_counter}
\title{FUNCTION_TITLE}
\usage{
get_counter(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
