% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{FUNCTION_TITLE}
\usage{
get_nodes(tag)
}
\arguments{
\item{tag}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xml2]{read_xml}},\code{\link[xml2]{xml_child}},\code{\link[xml2]{xml_find_all}},\code{\link[xml2]{xml_text}},\code{\link[xml2]{xml_double}}
}
