% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{FUNCTION_TITLE}
\usage{
get_time(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
