% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job2chunk.R
\name{job2chunk}
\alias{job2chunk}
\title{FUNCTION_TITLE}
\usage{
job2chunk(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
