% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibble_app.R
\name{qibble_app}
\alias{qibble_app}
\title{FUNCTION_TITLE}
\usage{
qibble_app(workDir, tag, viewer = shiny::paneViewer())
}
\arguments{
\item{workDir}{PARAM_DESCRIPTION}

\item{tag}{PARAM_DESCRIPTION}

\item{viewer}{PARAM_DESCRIPTION, Default: shiny::paneViewer()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
