% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibble.R
\name{qibble}
\alias{qibble}
\title{FUNCTION_TITLE}
\usage{
qibble(workDir = NULL, tag = NULL, qapply_log_tail = 1)
}
\arguments{
\item{workDir}{PARAM_DESCRIPTION, Default: NULL}

\item{tag}{PARAM_DESCRIPTION, Default: NULL}

\item{qapply_log_tail}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
