% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibble_app.R
\name{qibble_app}
\alias{qibble_app}
\title{FUNCTION_TITLE}
\usage{
qibble_app(workDir, tag, launch.browser = TRUE)
}
\arguments{
\item{workDir}{PARAM_DESCRIPTION}

\item{tag}{PARAM_DESCRIPTION}

\item{launch.browser}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
